% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Histogram.R
\name{behavior_hist}
\alias{behavior_hist}
\title{Create a histogram with behavioral counts}
\usage{
behavior_hist(
  data,
  x_axis,
  y_axis,
  color = NULL,
  position = NULL,
  width = NULL
)
}
\arguments{
\item{data}{the data being utilized- needs to be in total count form}

\item{x_axis}{the column that you want to use as the independent variable}

\item{y_axis}{the column that you want to use as the dependent variable}

\item{color}{what column you want the bars separated by}

\item{position}{what geom_col position you want the bars in}

\item{width}{the width of the bars}
}
\value{
a histogram of behavioral frequency
}
\description{
Uses the behavioral count transformed data set to visualize behavioral frequencies
between experimental conditions.
}
