% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lm.R
\name{linearmod}
\alias{linearmod}
\title{Linear modeling function}
\usage{
linearmod(data, dependent_col, independent_col)
}
\arguments{
\item{data}{the data set being utilized}

\item{dependent_col}{the response variable}

\item{independent_col}{the predictor variable}
}
\value{
the linear model summary
}
\description{
This function allows you to run a linear model on selected columns within a
data set.
}
