% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeparateCols.R
\name{separate_cols}
\alias{separate_cols}
\title{Separate experimental group columns}
\usage{
separate_cols(data, column, newname1, newname2)
}
\arguments{
\item{data}{the data set being utilized}

\item{column}{the column that is being split}

\item{newname1}{the desired name of the column created from the characters to the left of the delimiter aka _}

\item{newname2}{the desired name of the column created from the characters to the right of the delimiter aka _}
}
\value{
the two new separated columns
}
\description{
Uses the experimental group column to create new columns so sensory manipulations
can be isolated if desired
}
