% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StandardError.R
\name{stnd_e}
\alias{stnd_e}
\title{Calculate standard error}
\usage{
stnd_e(data, column)
}
\arguments{
\item{data}{the data set being utilized}

\item{column}{the column you want to calculate SE for}
}
\value{
the selected column name followed by the calculated SE
}
\description{
This function will calculate the standard error for a selected row.
}
